created simple character effects specifically instant character effects framework

+++ Dice Tower Defence/Assets/Scripts/Character/Character/CharacterStats.cs
@@ -4,9 +4,9 @@ namespace UB
 {
     public class CharacterStats : ScriptableObject
     {
-        public int CurrentHealth;
-        public int MaxHealth;
-        public int Damage;
+        public float CurrentHealth;
+        public float MaxHealth;
+        public float Damage;
         public float MovementSpeed;
         public float Acceleration;
         public float StoppingDistance;
